requirements.txt
discord
discord.py
python-dotenv
requests
anthropic
aiohttp
certifi
async-timeout 
platform